+++ src/app/sprite_picker.hpp
@@ -13,6 +13,6 @@ enum class pose : int
 };
 
 sf::IntRect pick_sprite(int size, int row, int col);
-sf::IntRect pick_sprite(int size, aco::pose pose, aco::dir dir, int frame);
+sf::IntRect pick_sprite(float size, aco::pose pose, aco::dir dir, int frame);
 
 }
